Warn when compile stops Play Mode without enabled pause points
Compile already warned when pause points were active, but Play-only
compiles returned no Warning even though the session was discarded.
Keep the existing pause-point wording unchanged.
