+++ Packages/src/Editor/FirstPartyTools/Compile/CompilePlayModeStopWarningBuilder.cs.meta
